wtsapi: Use a smaller array cast to allow compilation for 32-bit architectures
In EnumerateSessions a pointer is returned from the API call. The pointer is
reinterpreted as an array of arbitrary length, then sliced down to its actual
size. The arbitrary array length that was previously chosen resulted in an
array that was too large to fit in a 32-bit address space, given the 12 byte
size of the sessionEntry structures. This resulted in a compilation error
when GOARCH=386.

A smaller arbitrary length for the array allows the compilation to succeed.